refactor: Add MockV3Aggregator contract for testing
This commit adds the MockV3Aggregator contract, which is used for testing other contracts' ability to read data from an aggregator contract. It includes functions to update the answer and round data, as well as retrieve the latest and specific round data. This addition improves the testing capabilities and flexibility of the codebase.